dave_cloud_main/requirements.txt
alembic==1.15.2
aniso8601==10.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
distlib==0.3.9
et_xmlfile==2.0.0
filelock==3.17.0
Flask==3.1.0
flask-cors==5.0.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
httplib2==0.22.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.4.2
kagglehub==0.3.11
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
marshmallow-sqlalchemy==1.4.1
msal==1.32.0
oauthlib==3.2.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
platformdirs==4.3.6
proto-plus==1.26.1
protobuf==6.30.2
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pygame==2.6.1
PyJWT==2.10.1
pyparsing==3.2.3
pyspark==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
SQLAlchemy==2.0.40
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.13.1
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.29.1
Werkzeug==3.1.3
xgboost==3.0.0
